Removed dma::mux::shared module
